Thanks for suggestion

+++ src/mlpack/bindings/python/print_wrapper_py.cpp
@@ -64,7 +64,7 @@ void PrintWrapperPY(const std::string& category,
   }
 
   // Import different mlpack programs that are to be wrapped.
-  for(long unsigned int i=0; i<methods.size(); i++)
+  for(size_t i=0; i<methods.size(); i++)
   {
     cout << "from mlpack." << groupName << "_" << methods[i] << " ";
     cout << "import " << groupName << "_" << methods[i] << endl;
